split test methods into PipelineConfigSpec and TemplateEntryPointVariableSpec
